Moved image files into newly refactored hierarchy. Moved hierarchy into models directory. Fixed all (code) links accordingly.
